Remove URI encoded suffix from project name
If a uri encoded url was visited (eg. `/rails/rails%3Fpage%3D1`) then
the `?page=1` part got appended to the `name` of the project.

Thus a new project got added with name 'rails?page=1'. If it weren't for
the question mark the project wouldn't exist because it wouldn't be
detected on Github. But because of the '?', we didn't get a 404.

Fixed by removing suffixes starting with '?'.